feat(homepage): integrate genres and books with firebase
